Move components, upgrade typechain
