+++ package.json
@@ -17,7 +17,7 @@
     "postinstall": "yarn typechain && typesync",
     "lint": "next lint",
     "tsc": "tsc",
-    "typechain": "typechain --target ethers-v5 --outDir ./containers/Contracts ./containers/ABIs/**/*.json"
+    "typechain": "typechain --target=ethers-v5 --out-dir ./containers/Contracts ./containers/ABIs/**/*.json"
   },
   "dependencies": {
     "@clover-network/clover-connector": "^1.0.0",
@@ -95,7 +95,7 @@
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^11.1.0",
-    "@typechain/ethers-v5": "^6.0.0",
+    "@typechain/ethers-v5": "^10.0.0",
     "@types/chroma-js": "^2.1.3",
     "@types/coingecko-api": "^1.0.0",
     "@types/eslint": "^7.2.6",
@@ -128,7 +128,7 @@
     "postcss": "^8.4.5",
     "prettier": "2.1.2",
     "tailwindcss": "^3.0.15",
-    "typechain": "^4.0.0",
+    "typechain": "^8.0.0",
     "typescript": "^4.3.5",
     "typesync": "^0.8.0"
   },
