+++ v1/components/FarmsIntro.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 import { Trans, useTranslation } from "next-i18next";
 
-import { Connection } from "../containers/Connection";
-import { MiniIcon } from "../components/TokenIcon";
+import { Connection } from "../../containers/Connection";
+import { MiniIcon } from "./TokenIcon";
 import { noFarms, someFarms } from "v1/util/constants";
 import { ChainNetwork } from "picklefinance-core";
 import { Spacer } from "@geist-ui/react";
