+++ v1/components/JarPerformanceChart.js
@@ -1,4 +1,4 @@
-import { materialBlack, pickleGreen, cardColor } from "../v1/util/constants";
+import { materialBlack, pickleGreen, cardColor } from "../util/constants";
 import React, { useState, useEffect, useRef } from "react";
 import { LineChart, Line, YAxis, XAxis, Legend } from "recharts";
 import { makeStyles } from "@material-ui/core/styles";
