+++ v1/components/PercentageInput.tsx
