+++ v1/components/PickleIcon.tsx
