+++ v1/features/Balances/Balances.tsx
@@ -9,7 +9,7 @@ import { Connection } from "../../../containers/Connection";
 import { Prices as PriceComponent } from "../Prices/Prices";
 import { ethers } from "ethers";
 import { getProtocolData } from "v1/util/api";
-import PickleIcon from "../../../components/PickleIcon";
+import PickleIcon from "../../components/PickleIcon";
 
 const Container = styled(Grid.Container)`
   font-family: "Source Code Pro", sans-serif;
