+++ v1/features/DILL/Balances.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from "next-i18next";
 import { useBalances } from "../Balances/useBalances";
 import { DillStats } from "./DillStats";
 import { UseDillOutput } from "../../../containers/Dill";
-import PickleIcon from "../../../components/PickleIcon";
+import PickleIcon from "../../components/PickleIcon";
 
 const DataPoint = styled.div`
   font-size: 22px;
