+++ v1/features/DILL/Claim.tsx
@@ -8,7 +8,7 @@ import { Connection } from "../../../containers/Connection";
 import { UseDillOutput } from "../../../containers/Dill";
 import { ERC20Transfer } from "../../../containers/Erc20Transfer";
 import { formatPercent } from "../../util/number";
-import PickleIcon from "../../../components/PickleIcon";
+import PickleIcon from "../../components/PickleIcon";
 import { useButtonStatus, ButtonStatus } from "v1/hooks/useButtonStatus";
 
 const formatNumber = (num: number, precision?: number) =>
