+++ v1/features/MiniFarms/MiniFarmCollapsible.tsx
@@ -13,7 +13,7 @@ import { Jars } from "../../../containers/Jars";
 import { ERC20Transfer } from "../../../containers/Erc20Transfer";
 import Collapse from "../Collapsible/Collapse";
 import { JarApy } from "../../../containers/Jars/useJarsWithAPYPFCore";
-import { LpIcon, TokenIcon, MiniIcon } from "../../../components/TokenIcon";
+import { LpIcon, TokenIcon, MiniIcon } from "../../components/TokenIcon";
 import { useButtonStatus, ButtonStatus } from "v1/hooks/useButtonStatus";
 import { PickleCore } from "containers/Jars/usePickleCore";
