+++ v1/features/MiniFarms/MiniFarmList.tsx
@@ -11,7 +11,7 @@ import { JarCollapsible } from "./JarCollapsible";
 import { BalFarm } from "../PickleFarms/BalFarm";
 import { UniV3JarMiniFarmCollapsible } from "../UniV3/UniV3JarMiniFarmCollapsible";
 import { pickleWhite, someFarms } from "v1/util/constants";
-import { FarmsIntro } from "components/FarmsIntro";
+import { FarmsIntro } from "v1/components/FarmsIntro";
 import { PickleCore } from "containers/Jars/usePickleCore";
 import { useJarFarmMap } from "containers/Farms/farms";
 import { isJarDisabled, isJarActive } from "containers/Jars/jars";
